Created the player movement script and a Scene to test it. -Josiah
Updated the player movement and camera scripts so that the camera
will follow the player smoothly and the movement controls are relative
to the camera.  There is functionality to adjust the controllability
of the ball depending on if the ball is on the ground or not.

The test scene has a fun jump in it.  See if you can make the jump!