Nearly perfect the camera and movement. -Josiah
I went through and reworked a lot of the camera and solved a number
of bugs as well as added some different control features.  I removed
the player's ability to move backwards, now the S key is a very
powerful brake (it will need to be adjusted).  Also, I added a "dead
zone" where the camera will not adjust itself to the ball's velocity
vector.  This prevents a lot of jitters when the ball is still.  Within
this "dead zone" the player can rotate the camera without moving the ball
which is very helpful for when the player is trying to be careful.  There
is some jerkiness that occurs on occasion when the player moves out of
the "dead zone."  I also added a turn sensitivity and an option
to allow/disallow trichording.