Really cleaned up the camera controls. Adjusted inital and respwan camera glitches. -Josiah
There is certainly some crappy code in the Rolling Player controller and
camera follow scripts, however it gets the job done.  I will clean the code
up a bit more in the near future so that it isn't as inefficient, but it is
nearing the level of control that I want for this game.